add score, improve collision

+++ include/player.h
@@ -7,23 +7,23 @@ class Player
 {
     private:
         int id;
-        int x;
-        int y;
+        double x;
+        double y;
         int width;
         int height;
         double velocity;
         Texture* texture;
     public:
-        Player(int id, int x, int y);
-        Player(int id, int x, int y, int width, int height);
+        Player(int id, double x, double y);
+        Player(int id, double x, double y, int width, int height);
         void move();
         void draw();
-        void set_x(int x);
-        void set_y(int y);
+        void set_x(double x);
+        void set_y(double y);
         void set_velocity(double velocity);
         void set_texture(Texture* texture);
-        int get_x() const;
-        int get_y() const;
+        double get_x() const;
+        double get_y() const;
         int get_width() const;
         int get_height() const;
         double get_velocity() const;
